refactor: kakao SDK 대기 로직을 공용 유틸로 분리
